+++ config/namespaces.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * WP Plugin Info · Eliasis module for WordPress plugins
+ * WP Plugin Info · Eliasis plugin for WordPress plugins
  * 
  * @author     Josantonius - hello@josantonius.com
  * @copyright  Copyright (c) 2017
@@ -9,7 +9,7 @@
  * @since      1.0.0
  */
 
-$namespace = 'Eliasis\\Modules\\';
+$namespace = 'Eliasis\\Plugins\\';
 
 $module = 'WP_Plugin_Info\\Controller\\';