+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # CHANGELOG
 
+## 1.0.2 - 2017-09-09
+
+* Replaced `eliasis-framework/module` to `eliasis-framework/complement` library.
+
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Launcher\Launcher` class.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Launcher\Launcher->init()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Launcher\Launcher->admin()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Launcher\Launcher->afterAddMenu()` method.
+
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Admin\Info\Info` class.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Admin\Info\Info->getPluginsInfo()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Admin\Info\Info->getInfo()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Admin\Info\Info->updated()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Admin\Info\Info->getPluginInfo()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Controller\Admin\Info\Info->get()` method.
+
+* Deleted `App\Modules\WP_Plugin_Info\Model\Admin\Info\Info` class.
+* Deleted `App\Modules\WP_Plugin_Info\Model\Admin\Info\Info->getPluginsInfo()` method.
+* Deleted `App\Modules\WP_Plugin_Info\Model\Admin\Info\Info->setPluginsInfo()` method.
+
+* Added `App\Plugins\WP_Plugin_Info\Controller\Launcher\Launcher` class.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Launcher\Launcher->init()` method.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Launcher\Launcher->admin()` method.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Launcher\Launcher->afterAddMenu()` method.
+
+* Added `App\Plugins\WP_Plugin_Info\Controller\Admin\Info\Info` class.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Admin\Info\Info->getPluginsInfo()` method.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Admin\Info\Info->getInfo()` method.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Admin\Info\Info->updated()` method.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Admin\Info\Info->getPluginInfo()` method.
+* Added `App\Plugins\WP_Plugin_Info\Controller\Admin\Info\Info->get()` method.
+
+* Added `App\Plugins\WP_Plugin_Info\Model\Admin\Info\Info` class.
+* Added `App\Plugins\WP_Plugin_Info\Model\Admin\Info\Info->getPluginsInfo()` method.
+* Added `App\Plugins\WP_Plugin_Info\Model\Admin\Info\Info->setPluginsInfo()` method.
+
 ## 1.0.1 - 2017-09-06
 
 * Added `wp-plugin-info/wp_plugin-info.jsond` file.